backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
langchain
langchain-groq
psycopg2-binary
python-dotenv
alembic
langchain-google-genai
